# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: CIVORA API
  description: |
    CIVORA API - open digital infrastructure for public-interest service delivery.

    Version: 0.2.0-service-delivery

    ## Public-Interest Service Delivery Lifecycle

    Full end-to-end workflow for public service delivery:
    Request → Person registered → Case created (NEW) → OPEN →
    Eligibility assessed → Evidence collected → IN_REVIEW →
    Assessment completed → DECISION_PENDING → Decision made →
    APPROVED/REJECTED → IN_PROGRESS → Assistance provided →
    FOLLOW_UP → CLOSED

    State machine:
    NEW → OPEN → IN_REVIEW → ASSESSMENT → DECISION_PENDING →
    APPROVED → IN_PROGRESS → FOLLOW_UP → CLOSED
    REJECTED → CLOSED
  version: "0.2.0"
  contact:
    name: CIVORA
    url: https://github.com/alrazihi/civora
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local development server. Override with CIVORA_API_URL in production.
security: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT bearer token. The token must include `organization_id` claim.
        The API validates that the token's organization matches the `{orgId}`
        path parameter on all tenant-scoped endpoints.
...
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: INVALID_INPUT
            message:
              type: string
              example: "title is required"
...
    Assessment:
      type: object
      required: [id, organization_id, service_request_id, findings, recommendation, assessor, assessed_at, created_at]
      properties:
        id:
          type: string
          format: uuid
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        organization_id:
          type: string
          format: uuid
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        service_request_id:
          type: string
          format: uuid
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        findings:
          type: string
        needs_identified:
          type: string
          nullable: true
        recommendation:
          type: string
        assessor:
          type: string
          format: uuid
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        assessed_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    CreateAssessmentRequest:
      type: object
      required: [service_request_id, findings, recommendation]
      properties:
        service_request_id:
          type: string
          format: uuid
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        findings:
          type: string
        needs_identified:
          type: string
          nullable: true
        recommendation:
          type: string
...
    Assistance:
      type: object
      required: [id, organization_id, service_request_id, type, description, status, responsible_staff, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        organization_id:
          type: string
          format: uuid
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        service_request_id:
          type: string
          format: uuid
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        type:
          type: string
          enum: [FINANCIAL, FOOD, SHELTER, MEDICAL, EDUCATION, TRANSPORT, OTHER]
        description:
          type: string
        status:
          type: string
          enum: [PLANNED, IN_PROGRESS, COMPLETED, CANCELLED]
        responsible_staff:
          type: string
          format: uuid
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateAssistanceRequest:
      type: object
      required: [service_request_id, type, description, responsible_staff]
      properties:
        service_request_id:
          type: string
          format: uuid
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        type:
          type: string
          enum: [FINANCIAL, FOOD, SHELTER, MEDICAL, EDUCATION, TRANSPORT, OTHER]
        description:
          type: string
        responsible_staff:
          type: string
          format: uuid
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
    UpdateAssistanceStatusRequest:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [start, complete, cancel]
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /organizations/{orgId}/assessments:
    post:
      operationId: createOrgassessments
      tags: [Assessments]
      summary: Create an assessment
      description: Creates a needs assessment for a service request. Requires authentication.
      security:
        - BearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssessmentRequest'
      responses:
        '201':
          description: Assessment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assessment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /organizations/{orgId}/assessments/{assessmentId}:
    get:
      operationId: getAssessmentById
      tags: [Assessments]
      summary: Get an assessment
      description: Retrieves a specific assessment. Requires authentication.
      security:
        - BearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        - name: assessmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
      responses:
        '200':
          description: Assessment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assessment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /organizations/{orgId}/assessments/by-service-request/{serviceRequestId}:
    get:
      operationId: getAssessmentByServiceRequestId
      tags: [Assessments]
      summary: Get assessment by service request
      description: Retrieves the assessment for a specific service request. Requires authentication.
      security:
        - BearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        - name: serviceRequestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
      responses:
        '200':
          description: Assessment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assessment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /organizations/{orgId}/assistance:
    post:
      operationId: createOrgassistance
      tags: [Assistance]
      summary: Create an assistance action
      description: Creates a new assistance action for a service request. Requires authentication.
      security:
        - BearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssistanceRequest'
      responses:
        '201':
          description: Assistance created (status defaults to PLANNED)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assistance'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /organizations/{orgId}/assistance/{assistanceId}:
    get:
      operationId: getAssistanceById
      tags: [Assistance]
      summary: Get assistance
      description: Retrieves a specific assistance action. Requires authentication.
      security:
        - BearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        - name: assistanceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
      responses:
        '200':
          description: Assistance details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assistance'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /organizations/{orgId}/assistance/{assistanceId}/status:
    patch:
      operationId: updateAssistanceStatus
      tags: [Assistance]
      summary: Update assistance status
      description: Updates the status of an assistance action (start, complete, cancel). Requires admin or staff role.
      security:
        - BearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        - name: assistanceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAssistanceStatusRequest'
      responses:
        '200':
          description: Updated assistance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assistance'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Forbidden - admin or staff role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
  /organizations/{orgId}/assistance/by-service-request/{serviceRequestId}:
    get:
      operationId: listAssistanceByServiceRequestId
      tags: [Assistance]
      summary: List assistance by service request
      description: Lists assistance actions for a specific service request. Requires authentication.
      security:
        - BearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        - name: serviceRequestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 20
      responses:
        '200':
          description: List of assistance actions with pagination
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Assistance'
        '401':
          $ref: '#/components/responses/Unauthorized'
...